lits products in the page, Edit and almost the update working properly
Edit and almost the update working properly, So I am able to list all the products in the product page, I am also able to edit it with another form, so that I dont need to go to another page. And the update is almost working perfect, as it is bring all the data required.