Bug related to /login page fix and bug related to responsive button fix too
Bug related to /login page fix and bug related to responsive button fix too
The login gived a mistake because I did not put the hosting in a mode of single web application and also setup the redirection of all pages to index.html

About the responsive button, it was a bootstrap property that I needed to activate with code in the page.
Now it is working much better, but It is still missing some if statements for fixing the click when you move to a different location.