Doing the shopping cart. This is my first commit of the shopping cart service, and components
Doing the shopping cart. This is my first commit of the shopping cart service, and components.

Here I can finally create a shopping cart, for every user that is using the application. And after that I can take the product that he is adding to that shopping cart and I can add the quantity of products.

I can not delete products yet. But It will be my next step