Quantity for the navBar
I created a Quantity for the navBar, this allows me to see how many items are in the shoppingCart.
For this I created an interface that describe the properties of the objects that are coming from the database . Then I created a in the ngOnInit() a counting that returned all the quantities in the products items of the cart user.