I just for practiciting improved the database table structure
I just for practiciting improved the database table structure. Changing the models and classes that where modeling the data structure. Also changing the service where I was pushing the data to the database. And for ending I changed the HTML structure to make it consistence with the new strcutre of the data in my model.