Lot of improvementes and a new button
I improve the delete from the database, now when the item in the cart is 0, it gets delete from the database, and not only hide. Also I improved the html view with better space within the table and I created a button for deleting all the items in the shoppingCart of the user.