Last I will only take a look of the code and do some last improvements. But so far, everything is done.
Last I will only take a look of the code and do some last improvements.

I cleaned some constructors, as their code was able to be outside in methods in order to be more maintainable. It was making the code more clear an readable.
That was in products.component.ts
I also changed other code in the html as I added the async pipe with an observable in order to taking down a subscription.

I also changed the way I was making inicializaction of the shopping-cart-items.ts class, as I was adding its constructor. Now it can create new instances with a define structure.