In this commit, I created a form, and store the data from this form to the database.
In this commit, I created a form, and store the data from this form to the database.  The different this time, is that I am using addicional data from oder components and services and I am also storing it with the button of that form.

It was an interesting example. to store data from oder components too as I needed to map them as new objects and properties from this new object that I created.